Task2_fixed

+++ task2/task2_2.py
@@ -3,6 +3,6 @@ amountC = (input("and cents: "))
 amount = float(amountD + "." + amountC)
 quantity = int(input("Customer bought "))
 
-totalAm = float(amount * quantity)
+totalAm = amount * quantity
 
 print("Total is ", totalAm, "$")
